feat: add Postgres configuration to cmd (#21)
Add support for STORAGE=POSTGRES and related CLI options to the command.

Add a migrate subcommand to initialize Postgres databases.
