+++ internal/storage/storage_interface.go
@@ -62,6 +62,8 @@ type StorageProvider interface {
 	DeleteTransition(transitionID uuid.UUID) error
 	DeleteTransitionTask(transitionID uuid.UUID, taskID uuid.UUID) error
 	TASTransition(transition model.Transition, testVal model.Transition) (bool, error)
+	// Close closes the storage provider and releases any resources it holds.
+	Close() error
 }
 
 type DistributedLockProvider interface {
@@ -71,4 +73,6 @@ type DistributedLockProvider interface {
 	GetDuration() time.Duration
 	DistributedTimedLock(maxLockTime time.Duration) error
 	Unlock() error
+	// Close closes the distributed lock provider and releases any resources it holds.
+	Close() error
 }
